split to a new blk.go file
